requirements.txt
# Sera Otonom Dependencies

# Core
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async
aiohttp>=3.9.0
aiomqtt>=2.0.0
asyncio>=3.4.3

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0

# Data Processing
numpy>=1.26.0

# Utilities
python-dateutil>=2.8.2
tenacity>=8.2.0

# Logging
rich>=13.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
